+++ packages/transport/src/http-server/error.ts
